+++ trunk/pol-core/pol/core.h
@@ -38,44 +38,44 @@ bool move_character_to( Character* chr,
                        unsigned short x,
                         unsigned short y,
                         short z,
-                        long flags,
+                        int flags,
 						Realm* oldrealm );
 
-const long MOVEITEM_IGNOREMOVABLE	= 0x20000000L;
-const long MOVEITEM_FORCELOCATION	= 0x40000000L;
-const long RESURRECT_FORCELOCATION	= 0x01;
+const int MOVEITEM_IGNOREMOVABLE	= 0x20000000L;
+const int MOVEITEM_FORCELOCATION	= 0x40000000L;
+const int RESURRECT_FORCELOCATION	= 0x01;
 
-const long INSERT_ADD_ITEM          = 0x01;
-const long INSERT_INCREASE_STACK    = 0x02;
+const int INSERT_ADD_ITEM          = 0x01;
+const int INSERT_INCREASE_STACK    = 0x02;
 
-const long DELETE_BY_PLAYER         = 0x00;
-const long DELETE_BY_SCRIPT         = 0x01;
+const int DELETE_BY_PLAYER         = 0x00;
+const int DELETE_BY_SCRIPT         = 0x01;
 
-const long ENUMERATE_IGNORE_LOCKED	= 0x01;
-const long ENUMERATE_ROOT_ONLY		= 0x02;
+const int ENUMERATE_IGNORE_LOCKED	= 0x01;
+const int ENUMERATE_ROOT_ONLY		= 0x02;
 
-const long ITEMS_IGNORE_STATICS     = 0x01;
-const long ITEMS_IGNORE_MULTIS      = 0x02;
+const int ITEMS_IGNORE_STATICS     = 0x01;
+const int ITEMS_IGNORE_MULTIS      = 0x02;
 
-const long LIST_IGNORE_Z            = 0x40000000L;
+const int LIST_IGNORE_Z            = 0x40000000L;
 
-const long FP_IGNORE_MOBILES        = 0x01;
-const long FP_IGNORE_DOORS          = 0x02;
+const int FP_IGNORE_MOBILES        = 0x01;
+const int FP_IGNORE_DOORS          = 0x02;
 
-const long VENDOR_SEND_AOS_TOOLTIP  = 0x01;
-const long SENDDIALOGMENU_FORCE_OLD = 0x01;
+const int VENDOR_SEND_AOS_TOOLTIP  = 0x01;
+const int SENDDIALOGMENU_FORCE_OLD = 0x01;
 
-const long SAVE_FULL                = 0x00;
-const long SAVE_INCREMENTAL         = 0x01;
+const int SAVE_FULL                = 0x00;
+const int SAVE_INCREMENTAL         = 0x01;
 
-const long LOG_DATETIME             = 0x01;
+const int LOG_DATETIME             = 0x01;
 
-const long FINDSUBSTANCE_IGNORE_LOCKED = 0x01;
-const long FINDSUBSTANCE_ROOT_ONLY     = 0x02;
-const long FINDSUBSTANCE_FIND_ALL      = 0x04;
+const int FINDSUBSTANCE_IGNORE_LOCKED = 0x01;
+const int FINDSUBSTANCE_ROOT_ONLY     = 0x02;
+const int FINDSUBSTANCE_FIND_ALL      = 0x04;
 
-const long JOURNAL_PRINT_NAME			= 0x00;
-const long JOURNAL_PRINT_YOU_SEE		= 0x01;
+const int JOURNAL_PRINT_NAME			= 0x00;
+const int JOURNAL_PRINT_YOU_SEE		= 0x01;
 
 Item* find_walkon_item( Items& ivec, short z );
 void atomic_cout( const string& msg );
